requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
mysqlclient==2.2.4
packaging==24.0
PyJWT==1.7.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.6.0
